Upgrade to Qt 6
Qt 6 is widely available and Qt 6.2 has reached feature parity with Qt
5. It's time to switch.

qml/MainScreen.qml:
- Implicit signal arg passing is deprecated, use a lambda expression
  with a named argument instead.

qml/session/MetronomeMixer.qml:
- The Slider cannot bind to its background element's width/height
  because that creates a loop. Hardcode the same dimensions as the
  VerticalBar now. Maybe there is a nicer way to avoid duplicating the
  dimensions from VerticalBar.

core/SessionListModel.cpp:
- Replace Qt foreach() statement with C++ range-based for loop. This is
  now the preferred approach.

standalone/PortAudioEngine.cpp:
- The return type of QList::size() has changed from int to qsizetype.
  Explicitly cast to int.

vst/VstPlugin.cpp:
standalone/main.cpp:
- Qt::AA_EnableHighDpiScaling is the default and the attribute should no
  longer be set explicitly.

+++ standalone/PortAudioEngine.cpp
@@ -226,8 +226,8 @@ void PortAudioEngine::setInputRouting(const QList<ChannelRoute> &routing)
     }
     if (routing.size() != inputRouting_.size()) {
         qDebug("Got input routing with size %d, expected %d",
-               routing.size(),
-               inputRouting_.size());
+               static_cast<int>(routing.size()),
+               static_cast<int>(inputRouting_.size()));
         return;
     }
 
@@ -252,8 +252,8 @@ void PortAudioEngine::setOutputRouting(const QList<ChannelRoute> &routing)
     }
     if (routing.size() != outputRouting_.size()) {
         qDebug("Got output routing with size %d, expected %d",
-               routing.size(),
-               outputRouting_.size());
+               static_cast<int>(routing.size()),
+               static_cast<int>(outputRouting_.size()));
         return;
     }
 
@@ -516,7 +516,7 @@ bool PortAudioEngine::fillStreamParameters(PaStreamParameters *inputParams,
             if (inputRouting_.size() != params.channelCount) {
                 qCritical("Expected %d input channel routes, got %d",
                           params.channelCount,
-                          inputRouting_.size());
+                          static_cast<int>(inputRouting_.size()));
                 return false;
             }
 
@@ -531,7 +531,7 @@ bool PortAudioEngine::fillStreamParameters(PaStreamParameters *inputParams,
             if (outputRouting_.size() != params.channelCount) {
                 qCritical("Expected %d output channel routes, got %d",
                           params.channelCount,
-                          outputRouting_.size());
+                          static_cast<int>(outputRouting_.size()));
                 return false;
             }
